\name{DOSimEnv}
\alias{DOSimEnv}
\docType{data}
\title{
Disease Ontology enviroment object
}
\description{
	Disease Ontology enviroment object used for other functions
}
\usage{data(DOSimEnv)}

\source{
The original data is came from John D.Osborne's work.URL:\url{http://projects.bioinformatics.northwestern.edu/do_rif/}
}

\examples{
data(DOSimEnv)

ls(DOSimEnv)
}
\keyword{datasets}
