\name{getDefaultBackground}
\alias{getDefaultBackground}
\alias{DOGenes}

%- Also NEED an '\alias' for EACH other topic documented here.
\title{  Get a vector of genes annotated by current DO  }
\description{
   Get a vector of genes annotated by current DO . 
}
\usage{
	getDefaultBackground()
}
%- maybe also 'usage' for other objects documented here.

\value{
 	return a character vector of genes which are the genes annotated by current version
}
\author{ Jiang Li <\email{riverlee2008@gmail.com}> }
\keyword{file}

