\name{internal}
\alias{internal}
\alias{DOGraph}
%\alias{getDensityFactor}
%\alias{getDepthFactor}
\alias{getDisjAnc}
\alias{getDisjCommAncSim}
%\alias{getEnrichedSim}
%\alias{getGeneFeatures.internal}
\alias{getGSim}
%\alias{getWeightedDotSim}
\alias{precomputeTermSims}
\alias{initialize}
\alias{calcTermSim}
\alias{normalize.kernel}
\alias{getSimLch}
\alias{getSimPath}
\alias{getGIC}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{  internal functions  }
\description{
  internal functions: do not call these functions directly. 
}
\usage{
	various
}
%- maybe also 'usage' for other objects documented here.
\arguments{
	various
}
\value{
 	various
}
\author{ Jiang Li <\email{riverlee2008@gmail.com}> }
\keyword{file}

